Added validation

+++ controllers/cash_flow_controller.py
@@ -1,7 +1,6 @@
 from flask import Blueprint, request
 from datetime import date
 from init import db
-from models.debt import Debt, DebtSchema
 from models.cash_flow_items import CashFlowItem, CashFlowItemSchema
 from flask_jwt_extended import get_jwt_identity, jwt_required
 
@@ -50,31 +49,12 @@ def get_one_cash_flow_item(id):
     else:
         return {'error': 'Item does not exist for user'}
 
-@cash_flow_bp.route('/debts/')
-@jwt_required()
-def get_debts():
-    stmt = db.select(Debt)
-    debts = db.session.scalars(stmt)
-
-    debts_flow_id = DebtSchema(many=True, only=['cash_flow_item_id']).dump(debts)
-    flow_list_ids = [i['cash_flow_item_id'] for i in debts_flow_id]
-    print(flow_list_ids)
-
-    user = get_jwt_identity()
-    stmt = db.session.query(CashFlowItem).where(CashFlowItem.user_id == int(user))
-    expense = CashFlowItemSchema(many=True).dump(stmt)
-    user_debts = []
-    for i in flow_list_ids:
-        user_debts.append([item for item in expense if item.get('id') == i])
-
-    return {"User Debts": user_debts}
-
 # Create new entry in database for new item
 @cash_flow_bp.route('/<int:id>', methods=['POST'])
 @jwt_required()
 def create_item(id):
     data = CashFlowItemSchema().load(request.json)
-
+    # Create new instance of item
     cash_flow_item = CashFlowItem(
         description = data['description'],
         amount = data['amount'],
@@ -84,43 +64,41 @@ def create_item(id):
         category_id = id
     )
 
+    # Add and commit to database
     db.session.add(cash_flow_item)
     db.session.commit()
 
     return CashFlowItemSchema().dump(cash_flow_item), 201
 
 # Update current item in database if belogns to logged in user
-@cash_flow_bp.route('/<int:id>/', methods=['PUT', 'PATCH'])
+@cash_flow_bp.route('/<int:id>/', methods=['PUT', 'PATCH', 'DELETE'])
 @jwt_required()
 def update_item(id):
-    stmt = db.select(CashFlowItem).filter_by(id=id)
-    cash_flow_item = db.session.scalar(stmt)
-    user = get_jwt_identity()
-    if int(user) == int(cash_flow_item.user_id):
-        if cash_flow_item:
-            cash_flow_item.description = request.json.get('description') or cash_flow_item.description
-            cash_flow_item.amount = request.json.get('amount') or cash_flow_item.amount
-            cash_flow_item.frequency = request.json.get('frequency') or cash_flow_item.frequency
-            return CashFlowItemSchema(only=['description', 'amount', 'frequency']).dump(cash_flow_item)
-        else:
-            return {'error': f'Card not found with id {id}'}, 404
-    else:
-        return {'error': 'Card not yours!'}, 404
+    data = CashFlowItemSchema().load(request.json)
 
-# Delete item if belongs to logged in user
-@cash_flow_bp.route('/<int:id>/', methods=['DELETE'])
-@jwt_required()
-def delete_item(id):
     stmt = db.select(CashFlowItem).filter_by(id=id)
     cash_flow_item = db.session.scalar(stmt)
+    
     user = get_jwt_identity()
+    # Check if belongs to logged in user
     if int(user) == int(cash_flow_item.user_id):
-        if cash_flow_item:
-            db.session.delete(cash_flow_item)
-            db.session.commit()
-            return {'message': f"Item '{cash_flow_item.description}' deleted successfully"}
-        else:
-            return {'error': f'Item not found with id {id}'}, 404
+        if request.method == ('PUT' or 'PATCH'):
+            if cash_flow_item:
+                # Update required values in database
+                cash_flow_item.description = data['description'] or cash_flow_item.description
+                cash_flow_item.amount = data['amount'] or cash_flow_item.amount
+                cash_flow_item.frequency = data['frequency'] or cash_flow_item.frequency
+                return CashFlowItemSchema(only=['description', 'amount', 'frequency']).dump(cash_flow_item)
+            else:
+                return {'error': f'Item not found with id {id}'}, 404
+        elif request.method == 'DELETE':
+            if cash_flow_item:
+                # Delete item from database
+                db.session.delete(cash_flow_item)
+                db.session.commit()
+                return {'message': f"Item '{cash_flow_item.description}' deleted successfully"}
+            else:
+                return {'error': f'Item not found with id {id}'}, 404
     else:
         return {'error': 'Card not yours!'}, 404
